Backend/requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.1.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
